Select Button action added for PLAYER_CHAR GUI
I got the Select button to work with the combo box. In order for this to work properly, I changed how the Connection instance variable in Runner.java works. There is a new getter method that checks if the Connection is null or closed. If it is, it will create a new connection to the database before returning the connection.